refact: implementa context api no countdown
