+++ src/pages/Home/Countdown/index.tsx
@@ -1,8 +1,11 @@
-import { useEffect, useState } from 'react'
+import { useContext, useEffect, useState } from 'react'
 import { differenceInSeconds } from 'date-fns'
 import { CountdownContainer, Separator } from './styles'
+import { CyclesContext } from '..'
 
 export function Countdown() {
+  const { activeCycle, activeCycleId, markCurrentCycleAsFinished } =
+    useContext(CyclesContext)
   const [amountSecondsPassed, setAmountSecondsPassed] = useState(0)
 
   const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0
@@ -18,16 +21,7 @@ export function Countdown() {
         )
 
         if (diffSeconds >= totalSeconds) {
-          setCycles((state) =>
-            state.map((cycle) => {
-              if (cycle.id === activeCycleId) {
-                return { ...cycle, finishedDate: new Date() }
-              } else {
-                return cycle
-              }
-            }),
-          )
-          setActiveCycleId(null)
+          markCurrentCycleAsFinished()
           clearInterval(interval)
         } else {
           setAmountSecondsPassed(diffSeconds)
@@ -38,7 +32,22 @@ export function Countdown() {
     return () => {
       clearInterval(interval)
     }
-  }, [activeCycle, totalSeconds, activeCycleId])
+  }, [activeCycle, totalSeconds, activeCycleId, markCurrentCycleAsFinished])
+
+  const currentSeconds = activeCycle
+    ? totalSeconds - amountSecondsPassed
+    : totalSeconds
+
+  const minutesAmount = Math.floor(currentSeconds / 60)
+  const secondsAmount = currentSeconds % 60
+
+  const minutes = String(minutesAmount).padStart(2, '0')
+  const seconds = String(secondsAmount).padStart(2, '0')
+
+  useEffect(() => {
+    if (activeCycle) document.title = `${minutes}:${seconds}`
+    else document.title = `Ignite Timer`
+  }, [minutes, seconds, activeCycle])
 
   return (
     <CountdownContainer>
